refactor: put get_inline_data where it can be overridden
